Changed atribute in method
Now it's finding schoolkid by full name;
Also added validation for absence of schoolkid or chastisements

+++ remove_chastisements.py
@@ -1,11 +1,18 @@
 import argparse
 
 def remove_chastisements(schoolkid):
-    chastisement = Chastisement.objects.filter(schoolkid=schoolkid).get()
-    chastisement.delete()
+    schoolkid = Schoolkid.objects.filter(full_name__contains=schoolkid).first()
+    if not schoolkid:
+        print('Такой ученик не найден')
+    else:
+        chastisement = Chastisement.objects.filter(schoolkid=schoolkid).all()
+        if not chastisement:
+            print('Замечаний не найдено')
+        else:
+            chastisement.delete()
 	
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Удаление записей с замечаниями ученика')
-    parser.add_argument('schoolkid', help='Ученик, замечания которого следует удалить (в формате экземпляра модели Schoolkid)')
+    parser.add_argument('schoolkid', help='Ученик, замечания которого следует удалить (в формате строки)')
     args = parser.parse_args()
